backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
web3==6.11.3
web3storage==0.1.1
twilio==8.10.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
transformers==4.35.2
torch==2.1.1
huggingface-hub==0.19.4
exa-py==1.0.0
python-socketio==5.10.0
aiohttp==3.9.1
aiofiles==23.2.1
Pillow==10.1.0
openzeppelin-contracts==0.1.0
python-socketio[asyncio-client]==5.10.0

